Rename `layouts` directory & `RepoLinkLayout` [#40]
