+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { ErrorBoundary } from 'react-error-boundary';
 import ErrorFallbackPage from '@/pages/ErrorFallback';
 import LoginPage from '@/pages/Login';
 import SignupPage from '@/pages/Signup';
-import MainLayout from '@/Layouts/MainLayout';
+import MainLayout from '@/layouts/MainLayout';
 import FeedPage from '@/pages/Feed';
 import NotFoundPage from '@/pages/NotFound';
 import FriendsPage from '@/pages/Friends';
@@ -12,7 +12,7 @@ import AddFriendPage from '@/pages/Friends/Add';
 import FriendContentPage from '@/pages/Friends/FriendContent';
 import ProfilePage from '@/pages/Profile';
 import './App.css';
-import UniversalLayout from '@/Layouts/UniversalLayout';
+import RepoLinkLayout from '@/layouts/RepoLinkLayout';
 import HomePage from '@/pages/Home';
 
 const App = () => {
@@ -21,7 +21,7 @@ const App = () => {
       <ErrorBoundary FallbackComponent={ErrorFallbackPage}>
         <Routes>
           <Route path='/' element={<HomePage />} />
-          <Route element={<UniversalLayout />}>
+          <Route element={<RepoLinkLayout />}>
             <Route path='/login' element={<LoginPage />} />
             <Route path='/signup' element={<SignupPage />} />
             <Route element={<MainLayout />}>
